fix: Fix one user can apply a job posting multiple times issue

+++ src/main/java/com/springframework/CareerConnect/controllers/JobPostingController.java
@@ -3,9 +3,11 @@ package com.springframework.CareerConnect.controllers;
 import com.springframework.CareerConnect.Mapper.JobPostingMapper;
 import com.springframework.CareerConnect.Mapper.MapStructMapper;
 import com.springframework.CareerConnect.domain.JobPosting;
+import com.springframework.CareerConnect.exceptions.ResourceNotFoundException;
 import com.springframework.CareerConnect.model.JobPostingDTO;
 import com.springframework.CareerConnect.model.JobPostingRequest;
 import com.springframework.CareerConnect.services.JobPostingService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -79,9 +81,31 @@ public class JobPostingController {
 
 
     @DeleteMapping("companies/{companyId}/jobPosting/{jobId}")
-    public ResponseEntity<JobPostingDTO> deleteJobPosting(@PathVariable Long companyId, @PathVariable Long jobId) {
-        log.info("Deleting jobPosting with id {}", jobId);
-        jobPostingService.deleteJobPostingById(companyId,jobId);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<JobPostingDTO> deleteJobPosting(
+            @PathVariable Long companyId,
+            @PathVariable Long jobId,
+            HttpServletRequest request
+    ) {
+        if (companyId == null || jobId == null) {
+            log.error("CompanyId and JobId cannot be null");
+            throw new IllegalArgumentException("CompanyId and JobId cannot be null");
+        }
+
+        log.info("Deleting jobPosting with companyId {} and jobId {}", companyId, jobId);
+
+        try {
+            jobPostingService.deleteJobPostingById(companyId,jobId);
+            log.info("Deleted jobPosting with companyId {} and jobId {}", companyId, jobId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (ResourceNotFoundException e) {
+            log.error("Resource not found for companyId {} and jobId {}", companyId, jobId);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (IllegalArgumentException e) {
+            log.error("Illegal argument for delete jobPosting", e);
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } catch (Exception e) {
+            log.error("Unexpected error while deleting job posting", e);
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
